requirements.txt
boto3==1.26.27
pymongo==4.4.1
pandas==1.3.5
pyarrow==12.0.1
python-dotenv==0.21.1
mysql-connector-python==8.0.31